requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.16.0
pillow==11.2.1
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2025.2
